requirements.txt
python-telegram-bot>=21.0
Pillow>=10.0.0
piexif>=1.1.3
numpy>=1.24.0
scikit-image>=0.21.0
pytesseract>=0.3.10
python-dotenv>=1.0.0




